Fixed a bug in parseRequestMsg() which was causing incorrect logic throughout the rest of the parsing/message checking functions.
We now default to send 0 as code, "" for title, and "" for message,
if the function finds now faults in the format of the request.

We then check the actual command, code and protocol in checkMessages()

TODO: 2 bugs still exist in parsing, which should be fixed after the prototype stage.

Bug1: We detect we get an empty line, by searching for "\r\n\r\n".
Only one empty line needs to exist, and the logic works for making
sure only 1 exists... However if the message has a \n, or a \r
singularly, then we have no current method to check for this, and
error appropriately.

e.g: This request would be valid in the server program:
"POST HTTP/1.1\r\n\r\n#4242\n"

Bug2: The current functionality of the string to int converter would
not error if a string/char is found within the code. Instead it will
take all valid numbers before the occurance of an invalid char/string.

e.g. This request would be valid in the server program:
"POST HTTP/1.1\r\n\r\n#42a42"
The server would respond to this message, exactly the same, as if it
were to get this message:
"POST HTTP/1.1\r\n\r\n#42"
The latter being valid.

These should be fixed for the final version.